requirements.txt
pytz
numpy
pandas
openpyxl
Markdown
Pillow
scikit-image
openai==1.72.0
tiktoken
google-genai
dotenv
demjson3
pathvalidate
docstring_parser
sqlalchemy
pydantic-settings
jinja2
humps
composio
colorama
anthropic
httpx_sse
Markdown
rapidfuzz
rank-bm25
psutil
llama_index
llama-index-embeddings-google-genai
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydub
python-multipart
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-requests
SpeechRecognition
ffmpeg
pg8000
pgvector
json_repair
rich
psycopg2-binary
mcp
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
